Refactor CertRoutes MM-355
